Ignore Movements when the From stock is empty
This happens when movements are scheduled that logically cancel other
movements. For example, for a request, the buffer_backoff movement can
be added to the schedule before a send_to_replica is added. If that
buffer_backoff was scheduled to occur after the send_to_replica, it
will deplete the requests in the buffer stock that would otherwise
have been used by the buffer_backoff.

This commit makes these cases ignored.