Don't duplicate request movements from buffer
Previously in RequestsBufferedStock.Add(), if the buffer found active
replicas, it would schedule movement events to replicas for all requests
currently in its internal stock. But if no replicas were found, it would
instead schedule backoff events.

This behaviour was actually buggy. Suppose this timeline:

1. Request A arrives in buffer, finding no replicas. It is scheduled to
   backoff for 100ms.
2. Request A arrives again, gets backed off again for 100ms.
3. Request B arrives before Request A's backoff has expired. It finds
   that replicas are available. The buffer then loops over all waiting
   requests and schedules them to move to the replicas in 1ns.
4. Request A and Request B are moved to replicas.
5. Request B's scheduled backoff occurs, but Request B has already been
   moved.

The problem was in step 3: looping over all requests instead of dealing
with a single request per .Add().

On the other hand, this change introduces a different possibility:

1. Request A is scheduled to move to a replica.
2. Before Request A's movement occurs, the replica is terminated.
3. Request A is moved to the replica, which continues as usual, giving a
   misleading count of requests processed.

What will probably be necessary will be for terminated replicas to
return requests to the buffer, or for an OnSchedule mechanism to allow
updating logic, or event cancellation, or some mix of these.