Fix incorrect types in data store
SQLite does not enforce value types; column types in CREATE TABLE are
basically just hints. In previous code I assumed type checking and so
was putting strings into columns intended to be integers.

Because I didn't notice this error, I also did not update the schema
to account for the change in foreign key constraints.