Time-shift movements instead of ignoring them
Previously, movements with schedule collisions were placed in the
ignored movements list. This change allows the movement PQ to
search for the next empty time and shift the movement to that
time.

This also revealed a minor bug in how Environment validates the
range of permissible schedule times. It checked to see if a
given movement was scheduled after halt, but this check relied
on the coincidence that the schedule collision would reject
anything scheduled at halt after the Environment's own setup.
When the time-shifting was added this bug was exposed.

[pivotal/skenario#3]