Fix sinusoidal clamping on 60s
The proximate cause: I assumed that for a time.Time, the
time.Second() function would return a steadily increasing
value. But actually, it returns the modulus of a minute.

This meant that the value fed into the sinusoidal wave
formula was incorrectly limited and did not encompass
the full circular range.

The correct function is time.Unix(), which returns seconds
since epoch. This creates the expected cyclical behaviour.

The root cause is that I didn't write a test for this
traffic pattern. If I'd written a test like the one this
commit adds, I would probably have detected the problem
sooner.

[pivotal/skenario#19]