Split plots, showing correlated RPS
This commit has a number of related changes.

The first is to split the plots for replicas and requests. Shown
together with RPS, the original plot was becoming too busy to
interpret.

Doing this with the previous query was too hard, so the query logic
has been modified. Previously the approach to querying was to create
a pivoted table of running counts. This is easy for a human to read
but was a bit janky for vega-lite to interpret.

This commit changes the query approach into something more amenable
to vega-lite's data model.

[pivotal/skenario#20]