Removed Knative Buffering (#67)
Kubernetes HPA doesn't have request buffering like Knative does. So supporting HPA requires bypassing the buffering. And Knative buffering which has evolved significantly since Skenario was implemented. Knative buffering is quite complex with the activator behaving as both a revision-level buffer and a load balancer. And the concept of desired burst capacity changes how routing is configured. All this is too much to model in Skenario core. And modeling it external to Skenario in a plugin is outside the scope of this HPA project.

As such, we are removing the Knative request buffering.  Instead of requests landing in the requests_buffering stock, they now land in the requests_routing stock and immediately (next nanosecond) dispatched to an active replica in round-robin fashion.