Made simple CPU utilization model (#71)
Kubernetes HPA scales primarily on CPU utilization.  So we need to model CPU usage.  Rather than modelling actual CPU cycles, we model utilization based on the existing delay model.  And since most operations are not only CPU but also include some IO, we add additional delay for IO operations.

This change adds a CPU and IO model.  And parameters in the UI for setting the CPU and IO time required for each request (serial).

Algorithm

1. Calculate free cpu capacity.
2. Calculate how much cpu time we need to process this request.
3. Calculate how much time we need to process this request taking into account IO time
4. Calculate average cpu load for the request that is utilization for the request
5. Add this utilization to occupied cpu capacity
6. Calculate current utilization of cpu in percentage
7. Calculate delay by sakasegawaApproximation which plus processing time (which is calculated in the previous step) forms total time for processing a request